pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "iact_estimator"
authors = [
    { name = "Michele Peresano", email = "peresano.michele@gmail.com" },
    { name = "Julian Sitarek", email = "jsitarek@uni.lodz.pl" },
]
description = "Estimate the performance of an IACT telescope system to an observation."
readme = "README.md"
requires-python = ">=3.9"
keywords = ["iact", "cherenkov", "astronomy"]
license = { text = "BSD-3-Clause" }
classifiers = [
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Astronomy",
]
dependencies = [
    "astroplan",
    "astropy",
    "matplotlib",
    "gammapy",
    "seaborn",
    "starplot",
]
dynamic = ["version"]

[project.optional-dependencies]
docs = [
    "furo",
    "ipywidgets",
    "myst-nb",
    "numpydoc",
    "rstcheck[sphinx]",
    "sphinx-argparse",
    "sphinx-autoapi",
    "sphinx-copybutton",
    "sphinxcontrib-towncrier",
    "towncrier<24.7",
]
dev = ["pre-commit", "ruff", "iact_estimator[docs]", "iact_estimator[tests]"]
tests = ["pytest", "pytest-console-scripts"]

[project.scripts]
iact-estimator = "iact_estimator.scripts.main:main"

[tool.setuptools_scm]
write_to = "src/iact_estimator/_version.py"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yml", "*.txt", "*.ecsv"]

[tool.numpydoc_validation]
# https://numpydoc.readthedocs.io/en/latest/validation.html#built-in-validation-checks
checks = [
    "all",  # report on all checks, except the below
    "EX01",
    "SA01",
    "ES01",
    "RT02",
]
